+++ tests/testthat/test_grouped_ggcorrmat.R
@@ -5,7 +5,6 @@ context("grouped_ggcorrmat")
 testthat::test_that(
   desc = "grouped_ggcorrmat plots work",
   code = {
-
     testthat::skip_on_cran()
 
     # with grouping.var missing ---------------------------------------------